requirements.txt
Django==1.6.5
South==0.8.4
django-debug-toolbar==1.2
djangorestframework==2.3.13
sqlparse==0.1.11
wsgiref==0.1.2
